+++ R/resample_gifti.R
@@ -6,7 +6,7 @@
 #' @param target_fname Where to save the resampled file.
 #' @param hemisphere \code{"left"} (default) or \code{"right"}. An error will 
 #'  occur if the hemisphere indicated in the GIFTI metadata does not match.
-#' @param file_type \code{"metric"}, \code{"label"}, \code{"surface"}, or \code{NULL} 
+#' @param file_type \code{"metric"}, \code{"label"}, \code{"surf"}, or \code{NULL} 
 #'  (default) to infer from \code{original_fname}.
 #' @param original_res The resolution of the original file. If \code{NULL}
 #'  (default), infer from the file.