+++ R/resample_surf.R
@@ -1,9 +1,9 @@
-#' Resample a \code{"surface"} object
+#' Resample a \code{"surf"} object
 #'
-#' Resample a \code{"surface"} object by writing it to a GIFTI, using the Connectome
+#' Resample a \code{"surf"} object by writing it to a GIFTI, using the Connectome
 #'  Workbench to resample it, and then reading the new file.
 #'
-#' @param surf A \code{"surface"} object
+#' @param surf A \code{"surf"} object
 #' @param resamp_res The desired resolution
 #' @param hemisphere \code{"left"} or \code{"right"}. Only used if not indicated by 
 #'  \code{surf$hemisphere}. An error will be raised if it does not match the